+++ components/toolbar/components/avatar.js
@@ -1,4 +1,4 @@
-import Img from '../../img/img';
+import Img from '../../Img/Img';
 
 
 function Menu(props) {
