+++ pages/backend/restapi.mdx
@@ -3,7 +3,7 @@ title: User Management, Content Management, and Display with the REST API
 ---
 
 import { Callout } from 'nextra/components'
-import Img from '../../components/img/img'
+import Img from '../../components/Img/Img'
 
 
 # User Management, Content Management, and Display with the REST API
