+++ pages/frontend/embedding_api.mdx
@@ -5,7 +5,7 @@ title: Embedding Tableau Views
 import { Bleed } from 'nextra-theme-docs'
 import { Callout } from 'nextra/components'
 import Tableau from './../../components/tableau/tableau'
-import Img from '../../components/img/img'
+import Img from '../../components/Img/Img'
 
 # Embedding Tableau Views
